Adapt 'Add' modal to user groups
The new 'Add' modal should be
adapted to the User groups data.
