+++ src/components/MemberOf/MemberOfUserGroups.tsx
@@ -1,19 +1,16 @@
 import React from "react";
-// Repositories
-import { userGroupsInitialData } from "src/utils/data/GroupRepositories";
+// PatternFly
+import { Pagination, PaginationVariant } from "@patternfly/react-core";
 // Data types
 import { UserGroup } from "src/utils/datatypes/globalDataTypes";
+// Redux
+import { useAppSelector } from "src/store/hooks";
 // Components
 import MemberOfToolbarUserGroups, {
   MembershipDirection,
 } from "./MemberOfToolbar";
 import MemberOfUserGroupsTable from "./MemberOfTableUserGroups";
-import { Pagination, PaginationVariant } from "@patternfly/react-core";
-
-interface MemberOfUserGroupsProps {
-  showAddModal: () => void;
-  showDeleteModal: () => void;
-}
+import MemberOfAddModal from "./MemberOfAddModalUserGroups";
 
 function paginate<Type>(array: Type[], page: number, perPage: number): Type[] {
   const startIdx = (page - 1) * perPage;
@@ -21,7 +18,35 @@ function paginate<Type>(array: Type[], page: number, perPage: number): Type[] {
   return array.slice(startIdx, endIdx);
 }
 
+interface TypeWithName {
+  name: string;
+}
+
+// Filter functions to compare the available data with the data that
+//  the user is already member of. This is done to prevent duplicates
+//  (e.g: adding the same element twice).
+function filterUserGroupsData<Type extends TypeWithName>(
+  list1: Array<Type>,
+  list2: Array<Type>
+): Type[] {
+  // User groups
+  return list1.filter((item) => {
+    return !list2.some((itm) => {
+      return item.name === itm.name;
+    });
+  });
+}
+
+interface MemberOfUserGroupsProps {
+  usersGroupsFromUser: UserGroup[];
+  updateUsersGroupsFromUser: (newList: UserGroup[]) => void;
+}
+
 const MemberOfUserGroups = (props: MemberOfUserGroupsProps) => {
+  const userGroupsFullList = useAppSelector(
+    (state) => state.usergroups.userGroupList
+  );
+
   const [groupsNamesSelected, setGroupsNamesSelected] = React.useState<
     string[]
   >([]);
@@ -31,17 +56,22 @@ const MemberOfUserGroups = (props: MemberOfUserGroupsProps) => {
 
   const [searchValue, setSearchValue] = React.useState("");
 
-  const [usersGroupsFromUser] = React.useState<UserGroup[]>(
-    userGroupsInitialData
-  );
-
   const [membershipDirection, setMembershipDirection] =
     React.useState<MembershipDirection>("direct");
 
+  const [showAddModal, setShowAddModal] = React.useState(false);
+  const [showDeleteModal, setShowDeleteModal] = React.useState(false);
+
   // Computed "states"
   const someItemSelected = groupsNamesSelected.length > 0;
-  const shownUserGroups = paginate(usersGroupsFromUser, page, perPage);
-  const showTableRows = usersGroupsFromUser.length > 0;
+  const shownUserGroups = paginate(props.usersGroupsFromUser, page, perPage);
+  const showTableRows = props.usersGroupsFromUser.length > 0;
+
+  // Available data to be added as member of
+  const userGroupsFilteredData: UserGroup[] = filterUserGroupsData(
+    userGroupsFullList,
+    props.usersGroupsFromUser
+  );
 
   return (
     <>
@@ -50,14 +80,14 @@ const MemberOfUserGroups = (props: MemberOfUserGroupsProps) => {
         onSearchTextChange={setSearchValue}
         refreshButtonEnabled={true}
         deleteButtonEnabled={someItemSelected}
-        onDeleteButtonClick={props.showDeleteModal}
+        onDeleteButtonClick={() => setShowDeleteModal(true)}
         addButtonEnabled={true}
-        onAddButtonClick={props.showAddModal}
+        onAddButtonClick={() => setShowAddModal(true)}
         membershipDirectionEnabled={true}
         membershipDirection={membershipDirection}
         onMembershipDirectionChange={setMembershipDirection}
         helpIconEnabled={true}
-        totalItems={usersGroupsFromUser.length}
+        totalItems={props.usersGroupsFromUser.length}
         perPage={perPage}
         page={page}
         onPerPageChange={setPerPage}
@@ -71,7 +101,7 @@ const MemberOfUserGroups = (props: MemberOfUserGroupsProps) => {
       />
       <Pagination
         className="pf-v5-u-pb-0 pf-v5-u-pr-md"
-        itemCount={usersGroupsFromUser.length}
+        itemCount={props.usersGroupsFromUser.length}
         widgetId="pagination-options-menu-bottom"
         perPage={perPage}
         page={page}
@@ -79,6 +109,17 @@ const MemberOfUserGroups = (props: MemberOfUserGroupsProps) => {
         onSetPage={(_e, page) => setPage(page)}
         onPerPageSelect={(_e, perPage) => setPerPage(perPage)}
       />
+      {showAddModal && (
+        <MemberOfAddModal
+          showModal={showAddModal}
+          onCloseModal={() => setShowAddModal(false)}
+          availableData={userGroupsFilteredData}
+          groupRepository={props.usersGroupsFromUser}
+          updateGroupRepository={(newList: UserGroup[]) =>
+            props.updateUsersGroupsFromUser(newList)
+          }
+        />
+      )}
     </>
   );
 };
